Move getParentComment to state/comments
